requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
base58==2.1.1
bitarray==1.2.2
boto3==1.21.22
botocore==1.24.22
bracex==2.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
crytic-compile==0.2.2
cytoolz==0.11.2
docopt==0.6.2
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
Flask==2.0.3
frozenlist==1.3.0
ghp-import==2.0.2
hexbytes==0.2.2
idna==3.3
importlib-metadata==4.10.1
ipfshttpclient==0.8.0a2
itsdangerous==2.1.1
Jinja2==3.0.3
jmespath==1.0.0
jsonschema==3.2.0
livereload==2.6.3
lru-dict==1.1.7
Markdown==3.3.6
MarkupSafe==2.0.1
mergedeep==1.3.4
mkdocs==1.0.4
mkdocs-awesome-pages-plugin==2.1.0
mkdocs-material==4.0.1
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
packaging==21.3
parsimonious==0.8.1
pipreqs==0.4.11
prettytable==3.0.0
protobuf==3.19.4
pycryptodome==3.14.1
Pygments==2.11.2
pymdown-extensions==9.1
pyparsing==3.0.7
pyrsistent==0.18.1
pysha3==1.0.2
python-dateutil==2.8.2
python-decouple==3.6
PyYAML==6.0
pyyaml_env_tag==0.1
requests==2.27.1
rlp==2.0.1
s3transfer==0.5.2
six==1.16.0
slither-analyzer==0.8.2
solc-select==0.2.1
toolz==0.11.2
tornado==6.1
urllib3==1.26.9
uuid==1.30
varint==1.0.2
watchdog==2.1.6
wcmatch==8.3
wcwidth==0.2.5
web3==5.28.0
websockets==9.1
Werkzeug==2.0.3
yarg==0.1.9
yarl==1.7.2
zipp==3.7.0
